feat: rename to east_dunbartonshire, broaden scope beyond bins
