+++ custom_components/east_dunbartonshire/__init__.py
@@ -1,4 +1,4 @@
-"""The UK Bins integration."""
+"""The East Dunbartonshire integration."""
 
 from __future__ import annotations
 
@@ -7,13 +7,13 @@ from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant
 
 from .const import DOMAIN
-from .coordinator import UkBinsCoordinator
+from .coordinator import EastDunbartonshireCoordinator
 
 PLATFORMS: list[Platform] = [Platform.CALENDAR, Platform.SENSOR]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
-    coordinator = UkBinsCoordinator(hass, entry)
+    coordinator = EastDunbartonshireCoordinator(hass, entry)
     await coordinator.async_config_entry_first_refresh()
     hass.data.setdefault(DOMAIN, {})[entry.entry_id] = coordinator
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
