+++ custom_components/east_dunbartonshire/const.py
@@ -1,0 +1,12 @@
+"""Constants for the East Dunbartonshire integration."""
+
+DOMAIN = "east_dunbartonshire"
+
+CONF_UPRN = "uprn"
+CONF_ADDRESS = "address"
+
+BIN_TYPES = {
+    "food-caddy": "Food caddy",
+    "garden-bin": "Green bin",
+    "rubbish-bin": "Grey bin",
+}
