+++ custom_components/east_dunbartonshire/manifest.json
@@ -1,0 +1,13 @@
+{
+    "domain": "east_dunbartonshire",
+    "name": "East Dunbartonshire",
+    "codeowners": [
+        "@d0ugal"
+    ],
+    "config_flow": true,
+    "documentation": "https://github.com/d0ugal/east-dunbartonshire-bins",
+    "iot_class": "cloud_polling",
+    "issue_tracker": "https://github.com/d0ugal/east-dunbartonshire-bins/issues",
+    "requirements": [],
+    "version": "0.1.0"
+}
