--- custom_components/uk_bins/manifest.json
@@ -1,13 +1,0 @@
-{
-    "domain": "uk_bins",
-    "name": "UK Bins",
-    "codeowners": [
-        "@d0ugal"
-    ],
-    "config_flow": true,
-    "documentation": "https://github.com/d0ugal/uk-bins-homeassistant",
-    "iot_class": "cloud_polling",
-    "issue_tracker": "https://github.com/d0ugal/uk-bins-homeassistant/issues",
-    "requirements": [],
-    "version": "0.1.0"
-}
