+++ hacs.json
@@ -1,5 +1,5 @@
 {
-    "name": "UK Bins",
+    "name": "East Dunbartonshire",
     "render_readme": true,
     "content_in_root": false
 }
