+++ tests/integration/test_live_councils.py
@@ -9,7 +9,7 @@ Run with: uv run --extra integration pytest tests/integration/ -v
 
 import pytest
 
-from custom_components.uk_bins.coordinator import (
+from custom_components.east_dunbartonshire.coordinator import (
     _fetch_clackmannanshire,
     _fetch_east_dunbartonshire,
     _fetch_east_renfrewshire,
